% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_printsummaryrreviewobject.R
\name{print.summary.dataReviewRobject}
\alias{print.summary.dataReviewRobject}
\title{Printing summaryRreview Output}
\usage{
\method{print}{summary.dataReviewRobject}(x, ...)
}
\arguments{
\item{x}{an object of class "summary.dataReviewRobject", usually a result of a call to
\code{\link{summary.dataReviewRobject}}.}

\item{...}{Additional arguments passsed on to \code{\link{createTextSummary}}}
}
\description{
Printing summaryRreview Output
}
